requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
pyyaml>=6.0
scikit-learn>=1.3.0
umap-learn>=0.5.4
matplotlib>=3.7.0
seaborn>=0.12.0
wandb>=0.16.0
jupyter>=1.0.0
ipykernel>=6.0.0
tqdm>=4.65.0
